Removes local dependencies and updates workflow (#5)
## Overview
Removes the local JCuda dependencies contained within the `lib` directory and instead uses online dependencies. Also updates the build workflow to run on both push and pull to main, build with JDK 21, and adds a Docker build stage.

## Commits included
- Removed local JCuda packages
     - Removed local JCuda packages as they can be downloaded from the Maven
central repo in the build stage.
- Updates build workflow (#4)
     - Updates build workflow to run on both push & pull to main. Updated JDK
used for the build pipeline to 21 and adds a Docker build stage to build the
base image. Also renames the build workflow file to reflect that it is
no longer just a maven workflow.